exercicio 1 ao 18

+++ exercicios/sequencial/exercicio_8.py
@@ -1,4 +1,4 @@
-horas_trabalhadas = float(input("Digite o numero de horas trablhadas:"  ))
+horas_trabalhadas = int(input("Digite o numero de horas trablhadas:"  ))
 valor_horas = float(input("digite o valor da hora trabalhada:"))
 salario = horas_trabalhadas * valor_horas
 print(" o salario total é de",salario)
